+++ src/Model/Query.php
@@ -38,14 +38,28 @@ class Query
 	}
 
 
+	public static function get_terms($args=[])
+	{
+		$args['fields'] = 'ids';
+
+		$terms = get_terms( $args );
+
+		foreach ($terms as &$term)
+		{
+			$term = new Term( $term );
+		}
+
+		return $terms;
+	}
+
+
 	public static function get_posts($args=[], $fields=[])
 	{
 		global $wp_query;
 
-		if( !isset($args['post_type']) )
-			$args = array_merge($wp_query->query, $args);
+		$args = array_merge($wp_query->query, $args);
 
-		if( !isset($args['posts_per_page']))
+		if( !isset($args['posts_per_page']) and !isset($args['numberposts']))
 			$args['posts_per_page'] = get_option( 'posts_per_page' );
 
 		$args['fields'] = 'ids';